impléménetation des Dao et Manager + mise en place de la validation des Bean dans les Manager

+++ escalade/escalade-consumer/src/main/java/org/escalade/consumer/impl/dao/topo/TopoDaoImpl.java
@@ -6,6 +6,8 @@ import java.util.List;
 import javax.inject.Inject;
 import javax.inject.Named;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 import org.escalade.consumer.contract.dao.DaoFactory;
 import org.escalade.consumer.contract.dao.topo.TopoDao;
 import org.escalade.consumer.impl.dao.AbstractDaoImpl;
@@ -25,40 +27,52 @@ import org.springframework.jdbc.support.KeyHolder;
  */
 @Named("topoDao")
 public class TopoDaoImpl  extends AbstractDaoImpl implements TopoDao {
+	private static final Logger LOGGER = LogManager.getLogger(TopoDaoImpl.class);
 
 	@Inject
 	private RowMapper<Topo> topoRM;
 	@Inject
 	private DaoFactory daoFactory;
 
-	//READ
 	@Override
 	public Topo getTopo(String titre) {
-		String vSQL = "SELECT titre,description_id FROM public.topo WHERE titre = :titre";
-
-		MapSqlParameterSource vParams = new MapSqlParameterSource();
-		vParams.addValue("titre", titre);
-
-		NamedParameterJdbcTemplate vJdbcTemplate = new NamedParameterJdbcTemplate(getDataSource());
-
-		List<Topo> listTopo = vJdbcTemplate.query(vSQL, vParams, topoRM);
-
-		Topo topo;
-		if (listTopo.isEmpty()) {
-			topo = null;
-		} else {
-			topo = listTopo.get(0);
-			topo.setListPhotos(this.getListPhotos(topo.getTitre()));
-			topo.setListSpot(this.getListSpots(topo.getTitre()));
+		LOGGER.traceEntry("titre = " + titre);
+		
+		if(titre!=null && !titre.isEmpty()) {
+			String vSQL = "SELECT titre,description_id FROM public.topo WHERE titre = :titre";
+	
+			MapSqlParameterSource vParams = new MapSqlParameterSource();
+			vParams.addValue("titre", titre);
+	
+			NamedParameterJdbcTemplate vJdbcTemplate = new NamedParameterJdbcTemplate(getDataSource());
+	
+			List<Topo> listTopo = vJdbcTemplate.query(vSQL, vParams, topoRM);
+	
+			Topo topo;
+			if (listTopo.isEmpty()) {
+				topo = null;
+			} else {
+				topo = listTopo.get(0);
+				topo.setListPhotos(this.getListPhotos(topo.getTitre()));
+				topo.setListSpot(this.getListSpots(topo.getTitre()));
+			}
+			
+			LOGGER.traceExit(topo);
+			return topo;
+		}else {
+			LOGGER.traceExit("null");
+			return null;
 		}
-		return topo;
 	}
+	
 	/**
 	 * Retourne la liste des photos du topo à partir de son titre
 	 * @param titre
 	 * @return List<String>
 	 */
 	private List<String> getListPhotos(String titre) {
+		LOGGER.traceEntry("titre = " + titre);
+		
 		String vSQL = "SELECT nom_fichier FROM public.photo INNER JOIN public.photo_topo ON public.photo_topo.photo_id = public.photo.id WHERE titre_topo = :titreTopo" ;
 
 		MapSqlParameterSource vParams = new MapSqlParameterSource();
@@ -68,14 +82,18 @@ public class TopoDaoImpl  extends AbstractDaoImpl implements TopoDao {
 
 		List<String> listPhotos = vJdbcTemplate.queryForList(vSQL, vParams, String.class);
 
+		LOGGER.traceExit(listPhotos);
 		return listPhotos;
 	}
+	
 	/**
 	 * Retourne la liste des spots du topo à partir de son titre
 	 * @param titre
 	 * @return List<Spot>
 	 */
 	private List<Spot> getListSpots(String titre) {
+		LOGGER.traceEntry("titre = " + titre);
+		
 		String vSQL = "SELECT spot_id FROM public.spot_topo WHERE titre = :titre";
 
 		MapSqlParameterSource vParams = new MapSqlParameterSource();
@@ -87,17 +105,21 @@ public class TopoDaoImpl  extends AbstractDaoImpl implements TopoDao {
 		
 		List<Spot> listSpot = new ArrayList<Spot>();
 		
-		for (Integer integer : listSpotsId) {
-			Spot spot = daoFactory.getSpotDao().getSpot(integer.intValue());
-			listSpot.add(spot);
+		if(listSpotsId!=null) {
+			for (Integer integer : listSpotsId) {
+				Spot spot = daoFactory.getSpotDao().getSpot(integer.intValue());
+				listSpot.add(spot);
+			}
 		}
 		
+		LOGGER.traceExit(listSpot);
 		return listSpot;
 	}
 	
-	//CREATE
 	@Override
 	public void createTopo(Topo topo) {
+		LOGGER.traceEntry("topo = " + topo);
+		
 		if(topo!=null) {
 			ZoneTexte description = daoFactory.getZoneTexteDao().createZoneTexte(topo.getDescription());
 			
@@ -113,11 +135,18 @@ public class TopoDaoImpl  extends AbstractDaoImpl implements TopoDao {
 			
 			this.createPhotos(topo);
 			this.createListSpot(topo);
-			
 		}
+		
+		LOGGER.traceExit();
 	}
 	
+	/**
+	 * Enregistre les photos du topo dans la base deonnées
+	 * @param topo
+	 */
 	private void createPhotos(Topo topo) {
+		LOGGER.traceEntry("topo = " + topo);
+		
 		List<String> photos = topo.getListPhotos();
 		if(photos!=null) {
 			for (String photo : photos) {
@@ -143,8 +172,17 @@ public class TopoDaoImpl  extends AbstractDaoImpl implements TopoDao {
 				vJdbcTemplate2.update(vSQL2, vParams2);
 			}
 		}
+		
+		LOGGER.traceExit();
 	}
+	
+	/**
+	 * Enregistre la liste des spot du topo dans la base de données
+	 * @param topo
+	 */
 	private void createListSpot(Topo topo) {
+		LOGGER.traceEntry("topo = " + topo);
+		
 		List<Spot> spots = topo.getListSpot();
 		if(spots!=null) {
 			for (Spot spot : spots) {
@@ -160,31 +198,85 @@ public class TopoDaoImpl  extends AbstractDaoImpl implements TopoDao {
 				vJdbcTemplate.update(vSQL, vParams);
 			}
 		}
+		
+		LOGGER.traceExit();
 	}
 
-	//UPDATE
 	@Override
 	public void updateTopo(Topo topo) {
-		// TODO
+		LOGGER.traceEntry("topo = " + topo);
 		
+		daoFactory.getZoneTexteDao().updateZoneTexte(topo.getDescription());
+		
+		
+		this.deletePhotos(topo);
+		this.deleteListSpot(topo);
+		this.createPhotos(topo);
+		this.createListSpot(topo);
+		
+		LOGGER.traceExit();
 	}
 	
-	//DELETE
 	@Override
 	public void deleteTopo(String titre) {
-		ZoneTexte description = this.getTopo(titre).getDescription();
+		LOGGER.traceEntry("titre = " + titre);
 		
-		String vSQL = "DELETE FROM public.topo WHERE titre = :titre";
+		if(titre!=null && !titre.isEmpty()) {
+			ZoneTexte description = this.getTopo(titre).getDescription();
+			
+			//Rq : supprime en cascade les exemplaire de topo, liste des spot associé et les photos
+			String vSQL = "DELETE FROM public.topo WHERE titre = :titre";
+			
+			MapSqlParameterSource vParams = new MapSqlParameterSource();
+			vParams.addValue("titre", titre);
+			
+			NamedParameterJdbcTemplate vJdbcTemplate = new NamedParameterJdbcTemplate(getDataSource());
+			
+			vJdbcTemplate.update(vSQL, vParams);
+			
+			if(description != null) {
+				daoFactory.getZoneTexteDao().deleteZoneTexte(description.getId());
+			}
+		}
+		
+		LOGGER.traceExit();
+	}
+	
+	/**
+	 * Supprime de la base de données les photos associés au topo
+	 * @param topo
+	 */
+	private void deletePhotos(Topo topo) {
+		LOGGER.traceEntry("topo = " + topo);
+		
+		String vSQL = "DELETE FROM public.photo_topo WHERE titre_topo = :titreTopo";
 		
 		MapSqlParameterSource vParams = new MapSqlParameterSource();
-		vParams.addValue("titre", titre);
+		vParams.addValue("titreTopo", topo.getTitre());
 		
 		NamedParameterJdbcTemplate vJdbcTemplate = new NamedParameterJdbcTemplate(getDataSource());
 		
 		vJdbcTemplate.update(vSQL, vParams);
 		
-		if(description != null) {
-			daoFactory.getZoneTexteDao().deleteZoneTexte(description.getId());
-		}
+		LOGGER.traceExit();
+	}
+	
+	/**
+	 * Supprime de la base de données la listes des spots associés au topo
+	 * @param topo
+	 */
+	private void deleteListSpot(Topo topo) {
+		LOGGER.traceEntry("topo = " + topo);
+		
+		String vSQL = "DELETE FROM public.spot_topo WHERE titre = :titre";
+		
+		MapSqlParameterSource vParams = new MapSqlParameterSource();
+		vParams.addValue("titre", topo.getTitre());
+		
+		NamedParameterJdbcTemplate vJdbcTemplate = new NamedParameterJdbcTemplate(getDataSource());
+		
+		vJdbcTemplate.update(vSQL, vParams);
+		
+		LOGGER.traceExit();
 	}
 }
